refactor: rename schema to models
