+++ examples/postman_echo/request_methods/validate_with_functions_test.py
@@ -1,4 +1,4 @@
-# NOTICE: Generated By HttpRunner v3.0.8
+# NOTICE: Generated By HttpRunner v3.0.9
 # FROM: examples/postman_echo/request_methods/validate_with_functions.yml
 
 from httprunner import HttpRunner, Config, Step, RunRequest, RunTestCase
