+++ httprunner/loader.py
@@ -13,7 +13,7 @@ from pydantic import ValidationError
 
 from httprunner import builtin, utils
 from httprunner import exceptions
-from httprunner.schema import TestCase, ProjectMeta, TestSuite
+from httprunner.models import TestCase, ProjectMeta, TestSuite
 
 try:
     # PyYAML version >= 5.1
